jekyll script updated

+++ scripts/jekyll_setup.sh
@@ -1,6 +1,6 @@
 # ------------------------------------------------------------------------------------------
 #!/bin/bash
-# args: user_name
+# args: GitHub user_name
 # ------------------------------------------------------------------------------------------
 echo ${L1}
 echo "Setting up Jekyll for GitHub pages..."
@@ -8,6 +8,7 @@ echo ${L1}
 
 if [[ $# -ne 1 ]]; then
     echo "Illegal number of parameters" >&2
+    echo "./jekyll_setup.sh user_name (GitHub)"
     exit 2
 fi
 
@@ -19,13 +20,20 @@ sudo apt install -y ruby-full build-essential zlib1g-dev
 
 echo "Setting .bashrc"
 if [[ -z "${GEM_HOME}" ]]; then
-  echo 'export GEM_HOME="${HOME}/gems"' >> ~/.bashrc
-  echo 'export PATH="${HOME}/gems/bin:$PATH"' >> ~/.bashrc
+  # Load local .bashrc setttings
+  if [ -f ~/.bashrc_local ]; then
+    echo 'export GEM_HOME="${HOME}/gems"' >> ~/.bashrc_local
+    echo 'export PATH="${HOME}/gems/bin:$PATH"' >> ~/.bashrc_local
+  else
+    echo 'export GEM_HOME="${HOME}/gems"' >> ~/.bashrc
+    echo 'export PATH="${HOME}/gems/bin:$PATH"' >> ~/.bashrc
+  fi
   source ~/.bashrc
+  
 else 
   if [ "${GEM_HOME}" != "${HOME}/gems" ]; then
     echo "GEM_HOME wrong value, declared as: ${GEM_HOME}, expecting ${HOME}/gems}"
-    echo "Remove GEM_HOME declaration (maybe at .bashrc) and also in PATH"
+    echo "Remove GEM_HOME declaration (maybe at .bashrc or .bashrc_local) and also check PATH env.var."
     exit 1
   fi
 fi
@@ -45,8 +53,10 @@ cd ~/dev
 git clone git@github.com:${user}/${user}.github.io.git
 cd ${user}.github.io
 
-echo "Generating .gitignore for Jekyll"
-wget https://www.toptal.com/developers/gitignore/api/jekyll,python -O .gitignore
+if [ ! -f .gitignore ]; then
+  echo "Generating .gitignore for Jekyll"
+  wget https://www.toptal.com/developers/gitignore/api/jekyll,python -O .gitignore
+fi
 
 bundle init
 #bundle config set --local path 'vendor/bundle'
@@ -58,12 +68,10 @@ bundle install
 for f in *.markdown; do mv -- "$f" "${f%.markdown}.md"; done
 for f in **/*.markdown; do mv -- "$f" "${f%.markdown}.md"; done
 
-if [ -d "$DEVSETUP_DIR" ]; then
-  # Take action if $DIR exists. #
-  echo "Copying files from devsetup/jekyll ..."
-  cp ${DEVSETUP_DIR}/scripts/jekyll_files/* ./
-  
-  gem 'jekyll-seo-tag'
+if [ -f "./jekyll_setup_gems.lst" ]; then
+  while IFS= read -r line; do
+    gem ${line}
+  done < ./jekyll_setup_gems.lst
 fi
 
 #git add -A
@@ -78,5 +86,5 @@ fi
 #git checkout main
 
 echo ${L1}
-echo "Done! Happy coding..."
+echo "Done! Happy web coding..."
 echo ${L1}
